# docs/apifox-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "service-hub 接口文档",
    "description": "Service Hub 的 Agent 管理、命令下发与审计查询接口。",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/agents": {
      "get": {
        "tags": [
          "Agent 管理"
        ],
        "summary": "查询全部 Agent",
        "description": "返回当前所有 Agent 的最新连接状态与在线状态。",
        "operationId": "list_agents_api_agents_get",
        "responses": {
          "200": {
            "description": "请求成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AgentSnapshot"
                  },
                  "title": "Response List Agents Api Agents Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Agent 管理"
        ],
        "summary": "创建 Agent 并签发初始密钥",
        "description": "创建新的 Agent 记录，并返回该 Agent 的首次连接密钥。",
        "operationId": "provision_agent_api_agents_post",
        "parameters": [
          {
            "name": "X-Admin-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "管理令牌",
              "description": "Hub 管理令牌，值来自环境变量 ADMIN_TOKEN。"
            },
            "description": "Hub 管理令牌，值来自环境变量 ADMIN_TOKEN。"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentProvisionRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "请求成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentProvisionResponse"
                }
              }
            }
          },
          "422": {
            "description": "请求校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/agents/{agent_id}": {
      "get": {
        "tags": [
          "Agent 管理"
        ],
        "summary": "查询单个 Agent",
        "description": "根据 Agent 标识查询其最新状态。",
        "operationId": "get_agent_api_agents__agent_id__get",
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Agent 标识",
              "description": "要查询的 Agent 唯一标识。"
            },
            "description": "要查询的 Agent 唯一标识。"
          }
        ],
        "responses": {
          "200": {
            "description": "请求成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentSnapshot"
                }
              }
            }
          },
          "422": {
            "description": "请求校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/agents/{agent_id}/credentials/rotate": {
      "post": {
        "tags": [
          "Agent 管理"
        ],
        "summary": "轮换 Agent 密钥",
        "description": "为指定 Agent 重新签发连接密钥。",
        "operationId": "rotate_agent_credentials_api_agents__agent_id__credentials_rotate_post",
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Agent 标识",
              "description": "要轮换密钥的 Agent 唯一标识。"
            },
            "description": "要轮换密钥的 Agent 唯一标识。"
          },
          {
            "name": "X-Admin-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "管理令牌",
              "description": "Hub 管理令牌，值来自环境变量 ADMIN_TOKEN。"
            },
            "description": "Hub 管理令牌，值来自环境变量 ADMIN_TOKEN。"
          }
        ],
        "responses": {
          "200": {
            "description": "请求成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentCredentialResponse"
                }
              }
            }
          },
          "422": {
            "description": "请求校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/commands": {
      "get": {
        "tags": [
          "命令管理"
        ],
        "summary": "查询全局命令列表",
        "description": "分页查询所有 Agent 的命令历史，支持多条件筛选与排序。",
        "operationId": "list_commands_api_commands_get",
        "parameters": [
          {
            "name": "agentId",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Agent 标识",
              "description": "按 Agent 标识筛选。"
            },
            "description": "按 Agent 标识筛选。"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "状态",
              "description": "按命令状态筛选。"
            },
            "description": "按命令状态筛选。"
          },
          {
            "name": "action",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "动作",
              "description": "按命令动作筛选。"
            },
            "description": "按命令动作筛选。"
          },
          {
            "name": "requestedBy",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "请求发起方",
              "description": "按请求发起方筛选。"
            },
            "description": "按请求发起方筛选。"
          },
          {
            "name": "requestSource",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "请求来源",
              "description": "按请求来源筛选。"
            },
            "description": "按请求来源筛选。"
          },
          {
            "name": "createdAfter",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "起始时间",
              "description": "只返回创建时间大于等于该时间的命令。"
            },
            "description": "只返回创建时间大于等于该时间的命令。"
          },
          {
            "name": "createdBefore",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "结束时间",
              "description": "只返回创建时间小于等于该时间的命令。"
            },
            "description": "只返回创建时间小于等于该时间的命令。"
          },
          {
            "name": "sortBy",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(createdAt|updatedAt)$",
              "title": "排序字段",
              "description": "支持 createdAt 或 updatedAt。",
              "default": "createdAt"
            },
            "description": "支持 createdAt 或 updatedAt。"
          },
          {
            "name": "order",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(asc|desc)$",
              "title": "排序方向",
              "description": "支持 asc 或 desc。",
              "default": "desc"
            },
            "description": "支持 asc 或 desc。"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "title": "分页大小",
              "description": "单次返回的最大记录数。",
              "default": 100
            },
            "description": "单次返回的最大记录数。"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "title": "偏移量",
              "description": "分页偏移量。",
              "default": 0
            },
            "description": "分页偏移量。"
          }
        ],
        "responses": {
          "200": {
            "description": "请求成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandListResponse"
                }
              }
            }
          },
          "422": {
            "description": "请求校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/commands/{request_id}": {
      "get": {
        "tags": [
          "命令管理"
        ],
        "summary": "查询单条命令",
        "description": "根据请求 ID 查询单条命令的最新状态。",
        "operationId": "get_command_api_commands__request_id__get",
        "parameters": [
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "请求 ID",
              "description": "要查询的命令请求 ID。"
            },
            "description": "要查询的命令请求 ID。"
          }
        ],
        "responses": {
          "200": {
            "description": "请求成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandSnapshot"
                }
              }
            }
          },
          "422": {
            "description": "请求校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/commands/{request_id}/events": {
      "get": {
        "tags": [
          "命令管理"
        ],
        "summary": "查询命令事件",
        "description": "查询命令的完整审计事件流。",
        "operationId": "get_command_events_api_commands__request_id__events_get",
        "parameters": [
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "请求 ID",
              "description": "要查询事件流的命令请求 ID。"
            },
            "description": "要查询事件流的命令请求 ID。"
          }
        ],
        "responses": {
          "200": {
            "description": "请求成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CommandEventSnapshot"
                  },
                  "title": "Response Get Command Events Api Commands  Request Id  Events Get"
                }
              }
            }
          },
          "422": {
            "description": "请求校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AgentProvisionRequest": {
        "properties": {
          "agentId": {
            "type": "string",
            "title": "Agent 标识"
          }
        },
        "type": "object",
        "required": [
          "agentId"
        ],
        "title": "创建 Agent 请求"
      },
...
      "AgentSnapshot": {
        "properties": {
          "agentId": {
            "type": "string",
            "title": "Agent 标识"
          },
          "connected": {
            "type": "boolean",
            "title": "是否已连接"
          },
          "online": {
            "type": "boolean",
            "title": "是否在线"
          },
          "credentialConfigured": {
            "type": "boolean",
            "title": "是否已配置凭据"
          },
          "remote": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "远端地址"
          },
          "keyIssuedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "密钥签发时间"
          },
          "connectedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "连接时间"
          },
          "disconnectedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "断开时间"
          },
          "lastSeenAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "最后活跃时间"
          },
          "lastHeartbeatAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "最后心跳时间"
          },
          "lastPongAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "最后 Pong 时间"
          },
          "staleAfterSeconds": {
            "type": "integer",
            "title": "离线判定秒数"
          }
        },
        "type": "object",
        "required": [
          "agentId",
          "connected",
          "online",
          "credentialConfigured",
          "staleAfterSeconds"
        ],
        "title": "Agent 状态快照"
      },
...
    }
  },
...
}